+++ gui_engine/SaaA/js_init.c
@@ -2,8 +2,8 @@
 #include "js_user.h"
 extern int js_console_init(void);
 extern void jerry_port_set_default_context(jerry_context_t *context);
-extern void *gui_malloc(uint32_t n);
-extern void *gui_realloc(void *ptr_old, uint32_t n);
+extern void *gui_malloc(size_t n);
+extern void *gui_realloc(void *ptr_old, size_t n);
 extern void gui_free(void *rmem);
 jerry_value_t parsed_code;
 jerry_value_t ret_value;
