+++ gui_engine/server/gui_api.c
@@ -239,7 +239,7 @@ bool gui_mq_recv(void *handle, void *buffer, uint32_t size, uint32_t timeout)
 }
 
 
-void *gui_malloc(uint32_t n)
+void *gui_malloc(size_t n)
 {
     void *ptr = NULL;
     if ((n > os_api->mem_threshold_size) && (os_api->mem_threshold_size != 0))
@@ -264,7 +264,7 @@ void *gui_malloc(uint32_t n)
     return ptr;
 }
 
-void *gui_realloc(void *ptr_old, uint32_t n)
+void *gui_realloc(void *ptr_old, size_t n)
 {
     void *ptr = NULL;
 #ifdef ENABLE_RTK_GUI_OS_HEAP
@@ -355,7 +355,7 @@ void gui_mem_debug(void)
 }
 
 
-void *gui_lower_malloc(uint32_t n)
+void *gui_lower_malloc(size_t n)
 {
     void *ptr = NULL;
     GUI_ASSERT(lower_tlsf != NULL);
@@ -367,7 +367,7 @@ void *gui_lower_malloc(uint32_t n)
     return ptr;
 }
 
-void *gui_lower_realloc(void *ptr_old, uint32_t n)
+void *gui_lower_realloc(void *ptr_old, size_t n)
 {
     void *ptr = NULL;
     GUI_ASSERT(lower_tlsf != NULL);
