+++ gui_engine/server/gui_api.h
@@ -199,13 +199,13 @@ bool gui_mq_send_urgent(void *handle, void *buffer, uint32_t size, uint32_t time
 
 bool gui_mq_recv(void *handle, void *buffer, uint32_t size, uint32_t timeout);
 
-void *gui_malloc(uint32_t n);
-void *gui_realloc(void *ptr_old, uint32_t n);
+void *gui_malloc(size_t n);
+void *gui_realloc(void *ptr_old, size_t n);
 void gui_free(void *rmem);
 void gui_mem_debug(void);
 
-void *gui_lower_malloc(uint32_t n);
-void *gui_lower_realloc(void *ptr_old, uint32_t n);
+void *gui_lower_malloc(size_t n);
+void *gui_lower_realloc(void *ptr_old, size_t n);
 void gui_lower_free(void *rmem);
 void gui_lowr_mem_debug(void);
 void gui_sleep_cb(void);
